Project 3
Q5
Write a program that displays editable bar graphs. The GUI has two horizontal panels. The
right panel contains textfields with numbers. The left panel contains corresponding horizontal
bar graphs that graphically show the value of the numbers on the left. Use a rectangle with
the width (in pixels) equal to the numbers from the textfield, but no higher than 100 pixels.
The UI should look something like this:
Use MVC and the Observer pattern. Store the numbers in the model. Attach the graph view as a
listener. When the user edits one of the numbers, the number controller should update the model
and the model should notify the graph view that a changes has occurred. Then, the graph should get
the number from the model and repaint itself.
Hints: use method Textfield.getText() to extract the text from a textfield.
Add a key listener to each textfield with method
Textfield.addKeyListener(KeyListener l)

Q6
Write the Java code for classes Employee and Manager. Include a main() method in class Manager
where you illustrate how toJson() is used for objects of both classes.